Complete web fixes

+++ resources/views/tasks/index.blade.php
@@ -1,26 +1,34 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <title>Tasks</title>
-</head>
-<body>
-<h1>Task List</h1>
- <a href="{{ route('tasks.create') }}">
+@extends('layouts.base')
+
+@section('title', 'Task list')
+
+@section('content')
+@if (auth()->user()->hasVerifiedEmail())
+    <a href="{{ route('tasks.create') }}">
         <button>Create New Task</button>
     </a>
+@else
+    <p>Please verify your email address to start modifying the board.</p>
+    <form action="{{ route('verification.send') }}" method="POST">
+        @csrf
+        <button type="submit">Verify</button>
+    </form>
+@endif
 
 <ul>
     @foreach ($tasks as $task)
         <li>
             {{ $task->title }} - Assigned to: {{ $task->user->name ?? 'Unknown' }}
-            <a href="/tasks/{{ $task->id }}/edit">Edit</a> | <a href="/tasks/{{ $task->id }}/delete">Delete</a>
+            @if (auth()->user()->hasVerifiedEmail())
+                <a href="/tasks/{{ $task->id }}/edit">Edit</a> | <a href="/tasks/{{ $task->id }}/delete">Delete</a>
+            @endif
         </li>
     @endforeach
+
+    {{ $tasks->links('pagination::default') }}
 </ul>
 <form action="{{ route('logout') }}" method="POST" style="display: inline;">
     @csrf
     <button type="submit">Logout</button>
 </form>
-
-</body>
-</html>
+@endsection
